refactor for pytest errors
