+++ src/internal/view/modules/drag_board/__init__.py
@@ -1,10 +1,11 @@
+import internal.view.modules.modules
 import internal.view.dependencies
-import internal.view.modules.drag_board.states.drag_board
+import internal.view.modules.drag_board.states.drag_board_state
 
 
 def create_states(dependencies: internal.view.dependencies.Dependencies):
     dependencies.state_machine.add_state(
-        internal.view.modules.drag_board.states.drag_board.create_state(
+        internal.view.modules.drag_board.states.drag_board_state.create_state(
             dependencies.state_machine)
     )
